modifMessage Ok, likeMessage enlever le like NOK

+++ back/routes/messageCtrl.js
@@ -101,17 +101,27 @@ module.exports = {
         var userId = jwtUtils.getUserId(headerAuth);
 
         // Params
-        var title = req.body.title;
-        var content = req.body.content;
-        var attachement = `${req.protocol}://${req.get("host")}/images/${req.file.filename}`;
+        if (req.body.title) {
+            var title = req.body.title;
+            if (title.length < TITLE_LIMIT ) {
+                return res.status(400).json({ 'error' : 'Titre trop court '});
+            }
+        }
+        
+        if (req.body.content) {
+           var content = req.body.content;
+            if (content.length < CONTENT_LIMIT ) {
+                return res.status(400).json({ 'error' : 'Contenu trop court '});
+            }
+        }
 
-        if (title == null || content == null ) {
+        if (req.body.attachement) {
+            var attachement = `${req.protocol}://${req.get("host")}/images/${req.file.filename}`;
+        };
+
+        if (title === null && content === null && attachement === null) {
             return res.status(400).json({ 'error' : 'Paramétres manquants'});
-        }
-
-        if (title.length < TITLE_LIMIT || content.length < CONTENT_LIMIT ) {
-            return res.status(400).json({ 'error' : 'Titre ou contenu trop court '});
-        }
+        };
 
         var value = { title : title, content : content, attachement : attachement };
 
@@ -149,11 +159,16 @@ module.exports = {
         models.Message.findOne({where : {id : req.params.id }})
         .then(message => {
             if (userId === message.UserId || userIsAdmin === true ) {
-                const filename = sauce.imageUrl.split("/images/")[1];
-                fs.unlink(`images/${filename}`, () => {
+                if (message.attachement) {
+                    const filename = message.attachement.split("/images/")[1];
+                    fs.unlink(`images/${filename}`, () => {
+                        message.destroy()
+                        res.status(200).json({ message : " Message supprimé et attachement supprimé !"}); 
+                    })
+                } else {
                     message.destroy()
                     res.status(200).json({ message : " Message supprimé !"});
-                })  
+                }
             } else {
                 res.status(404).json({ 'error' : 'Utilisateur non autorisé'});
             }
@@ -168,36 +183,51 @@ module.exports = {
         var headerAuth = req.headers['authorization'];
         var userId = jwtUtils.getUserId(headerAuth);
         var likeStatus = req.body.like;
-
+        
         models.User.findOne({
             where: { id: userId }
-        })
+        })   
         .then(userFound =>{
             if (userFound) {
                 models.Message.findOne({ where : {id : req.params.id }})
                 .then(message => {
-                    if (likeStatus === 1) {
-                        models.Message.updateOne({ where : { id: req.params.id }},
-                            {$push: {usersLiked : userId}, $inc: {likes: +1 }})
-                        .then(() => res.status(200).json( {message: "Vous avez like ce message !" }))
-                        .catch(err => res.status(400).json({ err }));
-                    }
-                    if (likeStatus === 0) {
-                        const index = message.usersLiked.indexOf(userId);
-                        if (index > -1) {
-                            message.usersLiked.slice(index, 1);
-                            models.Message.updateOne({where : { id: req.params.id }},
-                                {$push: { usersLiked: {$each: [ ], $slice: index} }, $inc: { likes: -1 }})
-                            .then(() => res.status(200).json({ message: "Like annulé !" }))
-                            .catch((err) => res.status(400).json({ err }))
+                    if (likeStatus == 1) {
+                        var likes = message.likes + 1;
+                        if (message.usersLiked ) {
+                            var usersLiked = Array.from(message.usersLiked);
+                        } else {
+                            var usersLiked = [];
                         }
+                        usersLiked.push(userId)
+                        var value = { likes: likes, usersLiked: usersLiked};
+                        models.Message.update(
+                            value,
+                            { where : { id: req.params.id }}
+                        )
+                        .then(() => res.status(200).json( { message: "Vous avez like ce message !" }))
+                        .catch(err => res.status(400).json({ 'error' : "Vous n'avez pas pu liker ce message " + err }));
+                    }
+                    if (likeStatus == 0) {
+                        var likes = message.likes - 1;
+                        var usersLiked = Array.from(message.usersLiked);
+                        const index = usersLiked.indexOf(userId);
+                        usersLiked.slice(index, 1);
+                        var value = { likes: likes, usersLiked: usersLiked};
+                        models.Message.update(
+                            value,
+                            { where : { id: req.params.id }}
+                        )
+                        .then(() => res.status(200).json({ message: "Like annulé !" }))
+                        .catch((err) => res.status(400).json({ 'error' : "Vous n'avez pas pu annuler votre like "+ err }))
                     }
                 })
             } else {
+                console.log("6");
                 res.status(404).json({ 'error' : 'Utilisateur introuvable'});
             }
         })
         .catch(err => {
+            console.log("7");
             return res.status(409).json({ 'error' : "Impossible de vérifier l'utilisateur: " + err})
         });
     }
